Move import_handler module to testing.
We shouldn't make import_handler a publicly available module just
yet; we don't want to encourage other packages to import from it.
Since it's only used in testing, let's move this utility to
the testing package.
